Determine GIT_WORK_TREE using LOCAL_REPO
Using simply the REPO_NAME to determine the value for GIT_WORK_TREE required
that LOCAL_REPOS was stored directly under the home directory.  If it is
stored in a sub-directory of $HOME (e.g., ~/.local/movein), then the relative
path will be incorrect.

Instead, use the full path to the repo (LOCAL_REPO) as the basis for
determining how many directory traversals are required.

Signed-off-by: James McCoy <vega.james@gmail.com>